# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: WhyNoIPv6 API
  description: >
    The public read API of whynoipv6.com (07-api.md). Versionless root,
    snake_case wire, `{items,page,meta}` / `{points,meta}` envelopes, keyset
    cursor pagination, RFC 9457 problem+json errors. The only write path is
    the anonymous async live check (`POST /check`). Data is licensed
    CC-BY-NC-4.0.
  version: "1.0.0"
  contact:
    url: https://whynoipv6.com
  license:
    name: CC-BY-NC-4.0
    url: https://creativecommons.org/licenses/by-nc/4.0/
servers:
  - url: https://api.whynoipv6.com
# The filter composition every tier/scoped list shares with /domains
# (07 §4.4). YAML anchor — expanded per path at parse time.
...
paths:
...
  /check:
    post:
      operationId: createCheck
      tags: [check]
      summary: Enqueue an anonymous live check (async job + poll)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [host]
              properties:
                host: { type: string, example: example.com }
      responses:
        "202":
          description: Job enqueued; poll the Location header. Recommended poll interval 2 s.
          headers:
            Location:
              schema: { type: string, example: /check/481529 }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CheckAccepted" }
        "200":
          description: Dedupe hit (domain- or job-side) — a cached done envelope.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CheckEnvelope" }
        "400": { $ref: "#/components/responses/InvalidParameter" }
        "415": { $ref: "#/components/responses/UnsupportedMediaType" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /check/{id}:
    get:
      operationId: getCheck
      tags: [check]
      summary: Poll a live-check job to a terminal state
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, format: int64, minimum: 1 }
      responses:
        "200":
          description: >
            Terminal (`done`/`failed`) responses cache `public, max-age=60`;
            in-flight polls are `no-store`.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CheckEnvelope" }
        "404": { $ref: "#/components/responses/NotFound" }
  /badge/{host}.svg:
    get:
      operationId: getBadgeSVG
      tags: [badge]
      summary: Embeddable shields-flat status badge
      parameters: [{ $ref: "#/components/parameters/host" }]
      responses:
        "200":
          description: >
            Always 200 for a valid host (disabled/unknown render the gray
            `IPv6: unknown` variant). `Cache-Control: public, max-age=86400`.
          content:
            image/svg+xml:
              schema: { type: string }
        "400": { $ref: "#/components/responses/InvalidParameter" }
  /badge/{host}.json:
    get:
      operationId: getBadgeJSON
      tags: [badge]
      summary: shields.io endpoint-JSON badge variant
      parameters: [{ $ref: "#/components/parameters/host" }]
      responses:
        "200":
          description: >
            The shields.io endpoint schema — the one sanctioned camelCase
            exception. Embed via
            `https://img.shields.io/endpoint?url=https://api.whynoipv6.com/badge/example.com.json`.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ShieldsBadge" }
        "400": { $ref: "#/components/responses/InvalidParameter" }
...
components:
  parameters:
    host:
      name: host
      in: path
      required: true
      description: eTLD+1 or subdomain; canonicalized before lookup.
      schema: { type: string, example: example.com }
...
  responses:
...
    NotFound:
      description: Unknown or uncanonicalizable resource key.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    InvalidParameter:
      description: Malformed cursor, bad format, malformed badge host, bad window.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
...
    RateLimited:
      description: Over the per-/64 or global hourly quota; carries `retry_after` + Retry-After.
      headers:
        Retry-After:
          schema: { type: integer }
        RateLimit:
          schema: { type: string }
        RateLimit-Policy:
          schema: { type: string }
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/RateLimitProblem" }
    UnsupportedMediaType:
      description: The POST body is not JSON.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
...
  schemas:
    Classification:
      type: string
      enum: [unknown, inactive, sinner, partial, hero]
    IPv6Status:
      type: string
      enum: [supported, unsupported, no_record, not_applicable]
...
    ClassFlag:
      type: string
      enum: [broken_v6, www_missing, ns_missing, mail_missing, resources_v4only]
...
    Problem:
      type: object
      description: RFC 9457. `status` always equals the HTTP status line.
      required: [type, title, status]
      properties:
        type: { type: string, example: "https://whynoipv6.com/problems/not-found" }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
...
    RateLimitProblem:
      allOf:
        - $ref: "#/components/schemas/Problem"
        - type: object
          properties:
            retry_after: { type: integer, description: Seconds until the window frees. }
    StatusObject:
      type: object
      description: >
        Per-dimension confirmed status with provenance. `value` is null when
        the dimension has never been confirmed.
      required: [value, since]
      properties:
        value:
          allOf: [{ $ref: "#/components/schemas/IPv6Status" }]
          nullable: true
        since: { type: string, format: date-time, nullable: true }
    StatusBlock:
      type: object
      required: [base, www, ns, mx, conn, resources]
      properties:
        base: { $ref: "#/components/schemas/StatusObject" }
        www: { $ref: "#/components/schemas/StatusObject" }
        ns: { $ref: "#/components/schemas/StatusObject" }
        mx: { $ref: "#/components/schemas/StatusObject" }
        conn: { $ref: "#/components/schemas/StatusObject" }
        resources: { $ref: "#/components/schemas/StatusObject" }
...
    CheckAccepted:
      type: object
      description: Exactly these four keys on a 202.
      required: [id, host, status, created_at]
      properties:
        id: { type: integer, format: int64 }
        host: { type: string }
        status: { type: string, enum: [pending] }
        created_at: { type: string, format: date-time }
    CheckConfirmed:
      type: object
      required: [classification, class_flags, saint, status, as_of]
      properties:
        classification: { $ref: "#/components/schemas/Classification" }
        class_flags:
          type: array
          items: { $ref: "#/components/schemas/ClassFlag" }
        saint: { type: boolean }
        status: { $ref: "#/components/schemas/StatusBlock" }
        as_of: { type: string, format: date-time, nullable: true }
    CheckEnvelope:
      type: object
      description: >
        The §5.1.2 job body. `id` is null only on the domain-side dedupe
        envelope. `result` statuses are raw observations (incl.
        `inconsistent` on base/www), explicitly NOT confirmed state.
      required: [id, host, status, cached, created_at, completed_at, error, result, confirmed]
      properties:
        id: { type: integer, format: int64, nullable: true }
        host: { type: string }
        status: { type: string, enum: [pending, processing, done, failed] }
        cached: { type: boolean }
        created_at: { type: string, format: date-time }
        completed_at: { type: string, format: date-time, nullable: true }
        error: { type: string, nullable: true }
        result:
          type: object
          nullable: true
          additionalProperties: true
          properties:
            checked_at: { type: string, format: date-time }
            duration_ms: { type: integer, format: int64 }
            checks:
              type: object
              additionalProperties:
                type: object
                properties:
                  status: { type: string }
            latency:
              type: object
              properties:
                v4_ms: { type: integer, nullable: true }
                v6_ms: { type: integer, nullable: true }
        confirmed:
          allOf: [{ $ref: "#/components/schemas/CheckConfirmed" }]
          nullable: true
    ShieldsBadge:
      type: object
      description: shields.io endpoint schema — the one sanctioned camelCase exception.
      required: [schemaVersion, label, message, color, cacheSeconds, isError]
      properties:
        schemaVersion: { type: integer, enum: [1] }
        label: { type: string, enum: [IPv6] }
        message: { type: string, enum: [supported, full, partial, no IPv6, inactive, unknown] }
        color: { type: string, enum: [brightgreen, yellow, red, lightgrey] }
        cacheSeconds: { type: integer, example: 86400 }
        isError: { type: boolean }
...